ELK_Labs/Lab_2/training_pipeline/requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.15.0
matplotlib==3.7.2
seaborn==0.12.2
psutil==5.9.5
pyyaml==6.0.1